requirements.txt
absl-py==1.4.0
addict==2.4.0
aiohttp==3.8.3
aiosignal==1.3.1
albumentations==1.3.0
anyio==3.6.2
appdirs==1.4.4
argcomplete==2.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
Automat==22.10.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///opt/conda/conda-bld/beautifulsoup4_1650462163268/work
bleach==6.0.0
boto==2.49.0
brotlipy==0.7.0
cachetools==5.3.0
certifi @ file:///croot/certifi_1665076670883/work/certifi
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
chardet @ file:///home/builder/ci_310/chardet_1640804867535/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
coloredlogs==15.0.1
Comet==3.1.0
comm==0.1.2
conda==22.11.1
conda-build==3.23.3
conda-package-handling @ file:///croot/conda-package-handling_1666940373510/work
constantly==15.1.0
contourpy==1.0.7
coremltools==6.2
crcmod==1.7
cryptography @ file:///croot/cryptography_1665612644927/work
cycler==0.11.0
Cython==0.29.33
cython-bbox==0.1.3
debugpy==1.6.6
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml==0.7.1
dicomsdl==0.109.1
dnspython==2.2.1
docker-pycreds==0.4.0
et-xmlfile==1.1.0
exceptiongroup==1.0.4
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
exhaustive-weighted-random-sampler==0.0.2
expecttest==0.1.4
fasteners==0.18
fastjsonschema==2.16.2
filelock @ file:///opt/conda/conda-bld/filelock_1647002191454/work
flatbuffers==23.1.21
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
future==0.18.2
fvcore==0.1.5.post20221221
gast==0.4.0
gcs-oauth2-boto-plugin==3.0
gitdb==4.0.10
GitPython==3.1.30
glob2 @ file:///home/linux1/recipes/ci/glob2_1610991677669/work
google-apitools==0.5.32
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-reauth==0.1.1
grpcio==1.51.1
gsutil==5.20
htmlmin==0.1.12
httplib2==0.20.4
huggingface-hub==0.12.1
humanfriendly==10.0
hyperlink==21.0.0
hypothesis==6.61.0
idna @ file:///croot/idna_1666125576474/work
ImageHash==4.3.1
imageio==2.25.0
incremental==22.10.0
iopath==0.1.10
ipykernel==6.21.1
ipython @ file:///croot/ipython_1670919316550/work
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi @ file:///tmp/build/80754af9/jedi_1644315229345/work
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jstyleson==0.0.2
jupyter-events==0.6.3
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.2.1
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
libarchive-c @ file:///tmp/build/80754af9/python-libarchive-c_1617780486945/work
llvmlite==0.39.1
loguru==0.6.0
lxml==4.9.2
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mdurl==0.1.2
mistune==2.0.4
mkl-fft==1.3.1
mkl-random @ file:///home/builder/ci_310/mkl_random_1641843545607/work
mkl-service==2.4.0
monotonic==1.6
mpmath==1.2.1
multidict==6.0.4
multimethod==1.9.1
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==2.8.8
ninja==1.11.1
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1652801679809/work
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.8.0.121
nvidia-dali-nightly-cuda110==1.24.0.dev20230223
oauth2client==4.1.3
oauthlib==3.2.2
odfpy==1.4.1
onnx==1.13.0
onnx-graphsurgeon==0.3.26
onnx-simplifier==0.4.10
onnxoptimizer==0.3.8
onnxruntime==1.13.1
onnxruntime-gpu==1.14.1
onnxsim==0.4.17
opencv-python==4.7.0.68
opencv-python-headless==4.7.0.68
openpyxl==3.1.1
openvino==2022.3.0
openvino-dev==2022.3.0
openvino-telemetry==2022.3.0
packaging==23.0
pandas==1.3.5
pandas-profiling==3.6.6
pandocfilters==1.5.0
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pathtools==0.1.2
patsy==0.5.3
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
phik==0.12.3
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.3.0
pkginfo @ file:///croot/pkginfo_1666725041340/work
platformdirs==3.0.0
pluggy @ file:///tmp/build/80754af9/pluggy_1648024709248/work
portalocker==2.7.0
prometheus-client==0.16.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
protobuf==3.20.3
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pycosat @ file:///croot/pycosat_1666805502580/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.10.5
pydicom==2.3.1
Pygments==2.14.0
pylibjpeg==1.4.0
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.8.2
python-etcd==0.4.5
python-gdcm==3.0.21
python-json-logger==2.0.4
pytorch-ignite==0.4.11
pytz @ file:///opt/conda/conda-bld/pytz_1654762638606/work
pyu2f==0.1.5
PyWavelets==1.4.1
pyxlsb==1.0.10
PyYAML==6.0
pyzmq==25.0.0
qudida==0.0.4
regex==2022.10.31
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
requests-oauthlib==1.3.1
retry-decorator==1.1.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.1
rsa==4.7.2
ruamel.yaml @ file:///croot/ruamel.yaml_1666304550667/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1666302247304/work
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.9.3
seaborn==0.12.2
Send2Trash==1.8.0
sentry-sdk==1.16.0
setproctitle==1.3.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve @ file:///croot/soupsieve_1666296392845/work
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
statsmodels==0.13.5
sympy==1.11.1
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorrt==8.5.3.1
termcolor==2.2.0
terminado==0.17.1
texttable==1.6.7
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.2.3
tinycss2==1.2.1
tokenizers==0.13.2
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
toolz @ file:///croot/toolz_1667464077321/work
torch==1.13.1
torch2trt==0.4.0
torchelastic==0.2.2
torchtext==0.14.1
torchvision==0.14.1
tornado==6.2
tqdm @ file:///opt/conda/conda-bld/tqdm_1664392687731/work
traitlets @ file:///croot/traitlets_1671143879854/work
transformers==4.26.1
Twisted==22.10.0
typeguard==2.13.3
types-dataclasses==0.6.6
typing_extensions @ file:///croot/typing_extensions_1669924550328/work
uri-template==1.2.0
urllib3 @ file:///croot/urllib3_1670526988650/work
visions==0.7.5
wandb==0.13.10
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.2
widgetsnbextension==4.0.5
xlrd==2.0.1
yacs==0.1.8
yarl==1.8.2
ydata-profiling==4.0.0
-e git+https://github.com/Megvii-BaseDetection/YOLOX.git@618fd8c08b2bc5fac9ffbb19a3b7e039ea0d5b9a#egg=yolox
zope.interface==5.5.2
